Extracts methods to new class
