+++ tests/com/tkomiya/tests/VocabTest.java
@@ -1,20 +1,21 @@
 package com.tkomiya.tests;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
 
 import org.junit.Test;
 
 import com.tkomiya.models.Vocab;
-import com.tkomiya.testutils.TestVocabListLoader;
+import com.tkomiya.testutils.datageneration.vocab.VocabFactory;
 
 public class VocabTest {
 
 	@Test
 	public void testEquals() {
 		//Arrange
-		Vocab monday = TestVocabListLoader.mondayVocab();
-		Vocab monday2 = TestVocabListLoader.mondayVocab();
-		Vocab tuesday = TestVocabListLoader.tuesdayVocab();
+		Vocab monday = VocabFactory.mondayVocab();
+		Vocab monday2 = VocabFactory.mondayVocab();
+		Vocab tuesday = VocabFactory.tuesdayVocab();
 		
 		//Assert
 		assertEquals(monday, monday);
